pyproject.toml
[tool.poetry]
name = "a-novel-factory"
version = "0.1.0"
description = "An entry for NaNoGenMo 2022"
authors = ["Ben Steinberg"]
readme = "README.md"
packages = [{include = "a_novel_factory"}]
license = "GPL-3.0-or-later"

[tool.poetry.dependencies]
python = "^3.9"
factory-boy = "^3.2.1"
click = "^8.1.3"
humanize = "^4.4.0"

[tool.poetry.group.dev.dependencies]
flake8 = "^5.0.4"
pytest = "^7.2.0"

[tool.poetry.scripts]
new_novel = "a_novel_factory.main:new_novel"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
